+++ changelog.md
@@ -1,3 +1,10 @@
+# FluxLoader v0.10.0
+- Added `onLuaFilesLoaded` and `onLuaScriptExecute` events
+- Added `onChangeLanguage` events
+- Added translation system to `TranslationManager`
+- Fixed names of methods for exposing classes and objects
+- Fixed log output when creating backups and restoring
+
 # FluxLoader v0.9.0
 - Added methods for getting and setting lists of specific types for plugin configs
 - Added new methods for configuration files